feat: Refactor drawer funnel creating dynamic drawer and map it
