+++ src/views/insights/Dashboard.vue
@@ -11,22 +11,38 @@
         :style="getWidgetStyle(widget.grid_position)"
         :widget="widget"
         :isLoading="getWidgetLoadingStatus(widget)"
+        @open-update-config="openDrawerConfigWidget(widget)"
       />
     </template>
+    <DrawerConfigWidgetDynamic
+      v-show="!!widgetConfigurating"
+      :modelValue="showDrawerConfigWidget"
+      :widget="widgetConfigurating"
+      @close="showDrawerConfigWidget = false"
+    />
   </section>
 </template>
 
 <script>
 import { mapActions, mapState } from 'vuex';
 
-import DynamicWidget from '@/components/insights/widgets/DynamicWidget.vue';
 import { Dashboards } from '@/services/api';
+import DynamicWidget from '@/components/insights/widgets/DynamicWidget.vue';
+import DrawerConfigWidgetDynamic from '@/components/insights/drawers/DrawerConfigWidgetDynamic.vue';
 
 export default {
   name: 'DashboardView',
 
   components: {
     DynamicWidget,
+    DrawerConfigWidgetDynamic,
+  },
+
+  data() {
+    return {
+      showDrawerConfigWidget: false,
+      widgetConfigurating: null,
+    };
   },
 
   computed: {
@@ -110,6 +126,11 @@ export default {
       const config = widget.config;
       const isConfigured = config && Object.keys(config).length !== 0;
       return isConfigured ? !Object.keys(widget).includes('data') : false;
+    },
+
+    openDrawerConfigWidget(widget) {
+      this.widgetConfigurating = widget;
+      this.showDrawerConfigWidget = true;
     },
   },
 };
